+++ sell/src/main/java/com/chuanfangn/sell/entity/OrderMaster.java
@@ -9,6 +9,7 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import java.math.BigDecimal;
 
 @Data
 @DynamicUpdate
@@ -26,8 +27,8 @@ public class OrderMaster {
   private String buyerAddress;
   /**买家openid*/
   private String buyerOpenid;
-  /**`订单数量*/
-  private double orderAmount;
+  /**`订单总金额*/
+  private BigDecimal orderAmount;
   /**订单状态,默认为新下单*/
   private Integer orderStatus= OrderStatusEnums.NEW.getCode();
   /**支付状态,默认为0未支付*/
